// src/lib/api/generated/models/sourceMapNode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */
import type { UserId } from './userId';

export interface SourceMapNode {
  sellerId: UserId;
  /** @nullable */
  name: string | null;
  /** @nullable */
  lat: number | null;
  /** @nullable */
  lng: number | null;
  /** Total volume bought from this seller */
  totalVolumeOz: number;
  /** Total amount spent with this seller */
  totalSpend: number;
  /**
   * The most purchased produce type (useful for map icons)
   * @nullable
   */
  primaryProduceType: string | null;
  /** List of all produce types bought from this seller */
  produceCategories: string[];
}